+++ Tests/SwiftiomaticTests/Layout/ExpressionModifierTests.swift
@@ -102,8 +102,7 @@ struct ExpressionModifierTests: LayoutTesting {
       let b =
         await longerAsynchronousFunction()
       let c =
-        await
-        evenLongerAndLongerAsynchronousFunction()
+        await evenLongerAndLongerAsynchronousFunction()
 
       """